requirements.txt
# Telegram-related
telethon==1.34.0
cryptg==0.4.0
python-telegram==0.18.0

# Database
sqlalchemy==2.0.28
aiosqlite==0.19.0
alembic==1.13.1

# Web Framework
flask==3.0.2
werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.4
blinker==1.7.0

# Proxy Support
pysocks==1.7.1
aiohttp==3.9.3


# Utilities
python-dotenv==1.0.0
loguru==0.7.2
faker==22.6.0
colorama==0.4.6
typing-extensions>=4.9.0
pytz==2024.1

# Optional but recommended for better performance
ujson==5.9.0
uvicorn==0.27.1  # ASGI server for better performance
gunicorn==21.2.0  # Production WSGI server 